refactor: 완전한 공통 컴포넌트 외에는 common에서 components로 이동 #276
