+++ frontend/src/components/AlertModal/AlertModal.styled.ts
