+++ frontend/src/components/RoomSettingModal/CategoryDropdown/CategoryDropdown.styled.ts
