+++ frontend/src/components/RoomSettingModal/hooks/useCategoryDropdown.ts
