+++ frontend/src/components/RoomSettingModal/hooks/useCategoryListQuery.ts
