+++ frontend/src/components/layout/Header/Header.tsx
@@ -19,9 +19,9 @@ import useRoutePath from './hooks/useRoutePath';
 import ArrowLeft from '@/assets/images/arrowLeft.svg';
 import ExitIcon from '@/assets/images/exitIcon.svg';
 import SettingIcon from '@/assets/images/settingIcon.svg';
+import AlertModal from '@/components/AlertModal/AlertModal';
 import A11yOnly from '@/components/common/a11yOnly/A11yOnly';
-import AlertModal from '@/components/common/AlertModal/AlertModal';
-import RoomSettingModal from '@/components/common/RoomSettingModal/RoomSettingModal';
+import RoomSettingModal from '@/components/RoomSettingModal/RoomSettingModal';
 import useBalanceContentQuery from '@/hooks/useBalanceContentQuery';
 import useFocus from '@/hooks/useFocus';
 import useGetUserInfo from '@/hooks/useGetUserInfo';
